pyproject.toml
[project]
name = "personalityassesment"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "lightrag-hku>=1.1.5",
    "nano-vectordb>=0.0.4.3",
    "networkx>=3.4.2",
    "ollama>=0.4.7",
    "openai>=1.60.2",
    "pandas>=2.2.3",
    "pipmaster>=0.4.0",
    "pymupdf>=1.25.3",
    "python-dotenv>=1.0.1",
    "tenacity>=9.0.0",
    "textract>=1.6.5",
    "tiktoken>=0.9.0",
]
